⚒️ FIX : Se corrige novedad en flujo al eliminar recursos tanto en vimeo, youtube o gcore, no se estaba mostrando bien la informacion de feedback de las plataformas o no se mostraba

+++ src/Utilities/deleteFireStorageData.ts
@@ -2,20 +2,29 @@ import { fireStorage } from '@/helpers/firebase';
 
 export const deleteFireStorageData = async (fileUrl: string) => {
   let theFileWasDeleted = '';
-  // Create a reference to the file to delete
-  let fileRef = fireStorage.refFromURL(fileUrl);
+  try {
+    // Create a reference to the file to delete
+    let fileRef = fireStorage.refFromURL(fileUrl);
 
-  // Delete the file using the delete() method
-  await fileRef
-    .delete()
-    .then(function() {
-      // File deleted successfully
-      theFileWasDeleted = 'File deleted successfully';
-    })
-    .catch(function(error) {
-      // Some Error occurred
-      theFileWasDeleted = 'error deleting file';
-    });
+    // The reference is validated so as not to show a 404 by console in case a file is deleted and it is not saved and by chance the file is loaded and it is deleted again
+    const validatorTypeRef = fileRef;
+
+    if (typeof validatorTypeRef !== 'string') return 'the reference does not exist';
+
+    // Delete the file using the delete() method
+    await fileRef
+      .delete()
+      .then(function() {
+        // File deleted successfully
+        theFileWasDeleted = 'file deleted successfully';
+      })
+      .catch(function(error: any) {
+        // Some Error occurred
+        theFileWasDeleted = 'error deleting file';
+      });
+  } catch (error) {
+    theFileWasDeleted = 'unexpected error';
+  }
 
   return theFileWasDeleted;
 };
